+++ Project Envision/Models/Board/BoardModel.cs
@@ -87,5 +87,7 @@ namespace Project_Envision.Models
             set => m_gotusers = value;
         }
 
+
+
     }
 }
