+++ Project Envision/Views/Task/CreateTask.cshtml
@@ -1,15 +1,15 @@
 ﻿@model Project_Envision.Models.TaskPropertiesModel
-@{
-    ViewData["Title"] = "EditTask";
+@{ ViewData["Title"] = "CreateTask";
     Layout = "~/Views/Shared/_MainPageLayout.cshtml"; }
-}
-<link rel="stylesheet" href="~/css/EditTask.css" />
+
+<link rel="stylesheet" href="~/css/CreateBoard.css" />
 
 <div class="box" method="post">
     <form>
+
         <div>
             <label id="TaskNameLabel" asp-for="Task_name">Task Name:</label>
-            <input type="text" name="Task_name" asp-for="Task_name" value=@TaskPropertiesModel.getTask_name required="" />
+            <input type="text" name="Task_name" asp-for="Task_name" placeholder="Task Name" required="" />
 
         </div>
 
@@ -18,44 +18,43 @@
         </div>
 
         <br />
-        <label for="Task_Description" asp-for="Task_Description">Description:</label>
-        <br />
 
         <div>
-            <textarea id="Task_Description" name="Task_Description" rows="4" cols="25">@TaskPropertiesModel.getTask_Description</textarea>
+            <label for="Task_Description" asp-for="Task_Description">Description:</label>
+            <textarea cols="25" rows="4" placeholder="Task Description" id="description" asp-for="Task_Description" maxlength="100"></textarea>
         </div>
 
-        <br />
         <div>
             <span asp-validation-for="Task_Description" class="text-danger"></span>
         </div>
 
+        <br />
+
         <div>
             <label id="TaskPointsLabel" asp-for="Task_points">Task Points:</label>
-            <input type="number" min="0" max="100" name="Task_points" asp-for="Task_points" value=@TaskPropertiesModel.getTask_points required="" />
-
+            <input type="number" min="0" max="100" name="Task_points" asp-for="Task_points" placeholder="Task Points" required="" />
         </div>
-        <div>
 
+        <div>
             <span asp-validation-for="Task_points" class="text-danger"></span>
         </div>
 
         <br />
+
         <label for="Assignee" asp-for="Assignee">Choose a Assignee:</label>
 
         <select name="Assignee" id="Assignee" asp-for="Assignee">
 
             @{
-                <option value=@TaskPropertiesModel.getAssignee> @TaskPropertiesModel.getAssignee </option>
                 
                 if (TaskPropertiesModel.getAssignee != "None")
                 {
                     <option value="None"> None</option>
                 }
-
+                
                 for (int i = 0; i < BoardModel.m_usernamelist.Count(); i++)
                 {
-                    if (BoardModel.m_usernamelist[i] != TaskPropertiesModel.getAssignee && BoardModel.m_usernamelist[i] != null)
+                    if (BoardModel.m_usernamelist[i] != null)
                     {
 
                         <option value=@BoardModel.m_usernamelist[i]> @BoardModel.m_usernamelist[i]</option>
@@ -65,8 +64,8 @@
         </select>
 
         <div class="buttons">
-            <input type="submit" class="button homeButtons" value="Edit Task" />
-            <input type="button" class="button homeButtons" value="Cancel" onclick="location.href='@Url.Action("Board")'" />
+            <input type="submit" class="button homeButtons" value="Create Task" />
+            <input type="button" class="button homeButtons" value="Cancel" onclick="location.href='@Url.Action("ProductBacklog", "Board")'" />
         </div>
     </form>
 </div>
